+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/discount_code/DiscountCodeDraftBuilder.java
@@ -64,6 +64,7 @@ public class DiscountCodeDraftBuilder implements Builder<DiscountCodeDraft> {
 
     /**
      *  <p>User-defined unique identifier for the DiscountCode.</p>
+     *  <p>This field is optional for backwards compatibility reasons, but we strongly recommend setting it. Keys are mandatory for importing Discount Codes with the Import API and the Merchant Center.</p>
      * @param key value to be set
      * @return Builder
      */
@@ -391,6 +392,7 @@ public class DiscountCodeDraftBuilder implements Builder<DiscountCodeDraft> {
 
     /**
      *  <p>User-defined unique identifier for the DiscountCode.</p>
+     *  <p>This field is optional for backwards compatibility reasons, but we strongly recommend setting it. Keys are mandatory for importing Discount Codes with the Import API and the Merchant Center.</p>
      * @return key
      */