+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/inventory/InventoryEntryDraft.java
@@ -47,6 +47,7 @@ public interface InventoryEntryDraft extends com.commercetools.api.models.Custom
 
     /**
      *  <p>User-defined unique identifier for the InventoryEntry.</p>
+     *  <p>This field is optional for backwards compatibility reasons, but we strongly recommend setting it. Keys are mandatory for importing InventoryEntries with the Import API and the Merchant Center.</p>
      * @return key
      */
 
@@ -102,6 +103,7 @@ public interface InventoryEntryDraft extends com.commercetools.api.models.Custom
 
     /**
      *  <p>User-defined unique identifier for the InventoryEntry.</p>
+     *  <p>This field is optional for backwards compatibility reasons, but we strongly recommend setting it. Keys are mandatory for importing InventoryEntries with the Import API and the Merchant Center.</p>
      * @param key value to be set
      */