+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/inventory/InventoryEntryDraftImpl.java
@@ -71,6 +71,7 @@ public class InventoryEntryDraftImpl implements InventoryEntryDraft, ModelBase {
 
     /**
      *  <p>User-defined unique identifier for the InventoryEntry.</p>
+     *  <p>This field is optional for backwards compatibility reasons, but we strongly recommend setting it. Keys are mandatory for importing InventoryEntries with the Import API and the Merchant Center.</p>
      */
 
     public String getKey() {