+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/product/ProductData.java
@@ -84,7 +84,7 @@ public interface ProductData extends ProductDataLike {
     public LocalizedString getSlug();
 
     /**
-     *  <p>Title of the Product displayed in search results.</p>
+     *  <p>Title of the Product as displayed in search engine results.</p>
      * @return metaTitle
      */
     @Valid
@@ -92,7 +92,7 @@ public interface ProductData extends ProductDataLike {
     public LocalizedString getMetaTitle();
 
     /**
-     *  <p>Description of the Product displayed in search results below the meta title.</p>
+     *  <p>Description of the Product as displayed in search engine results below the meta title.</p>
      * @return metaDescription
      */
     @Valid
@@ -178,14 +178,14 @@ public interface ProductData extends ProductDataLike {
     public void setSlug(final LocalizedString slug);
 
     /**
-     *  <p>Title of the Product displayed in search results.</p>
+     *  <p>Title of the Product as displayed in search engine results.</p>
      * @param metaTitle value to be set
      */
 
     public void setMetaTitle(final LocalizedString metaTitle);
 
     /**
-     *  <p>Description of the Product displayed in search results below the meta title.</p>
+     *  <p>Description of the Product as displayed in search engine results below the meta title.</p>
      * @param metaDescription value to be set
      */