+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/product/ProductDraftBuilder.java
@@ -185,7 +185,8 @@ public class ProductDraftBuilder implements Builder<ProductDraft> {
 
     /**
      *  <p>User-defined unique identifier for the Product.</p>
-     *  <p>To update a Product using the Import API, the Product <code>key</code> must match the pattern <code>^[A-Za-z0-9_-]{2,256}$</code>.</p>
+     *  <p>This field is optional for backwards compatibility reasons, but we strongly recommend setting it. Keys are mandatory for importing Products with the Import API and the Merchant Center.</p>
+     *  <p>To update a Product using the Import API or Merchant Center, the Product <code>key</code> must match the pattern <code>^[A-Za-z0-9_-]{2,256}$</code>.</p>
      * @param key value to be set
      * @return Builder
      */
@@ -361,7 +362,7 @@ public class ProductDraftBuilder implements Builder<ProductDraft> {
     }
 
     /**
-     *  <p>Title of the Product displayed in search results.</p>
+     *  <p>Title of the Product as displayed in search engine results.</p>
      * @param builder function to build the metaTitle value
      * @return Builder
      */
@@ -373,7 +374,7 @@ public class ProductDraftBuilder implements Builder<ProductDraft> {
     }
 
     /**
-     *  <p>Title of the Product displayed in search results.</p>
+     *  <p>Title of the Product as displayed in search engine results.</p>
      * @param builder function to build the metaTitle value
      * @return Builder
      */
@@ -385,7 +386,7 @@ public class ProductDraftBuilder implements Builder<ProductDraft> {
     }
 
     /**
-     *  <p>Title of the Product displayed in search results.</p>
+     *  <p>Title of the Product as displayed in search engine results.</p>
      * @param metaTitle value to be set
      * @return Builder
      */
@@ -397,7 +398,7 @@ public class ProductDraftBuilder implements Builder<ProductDraft> {
     }
 
     /**
-     *  <p>Description of the Product displayed in search results.</p>
+     *  <p>Description of the Product as displayed in search engine results.</p>
      * @param builder function to build the metaDescription value
      * @return Builder
      */
@@ -409,7 +410,7 @@ public class ProductDraftBuilder implements Builder<ProductDraft> {
     }
 
     /**
-     *  <p>Description of the Product displayed in search results.</p>
+     *  <p>Description of the Product as displayed in search engine results.</p>
      * @param builder function to build the metaDescription value
      * @return Builder
      */
@@ -421,7 +422,7 @@ public class ProductDraftBuilder implements Builder<ProductDraft> {
     }
 
     /**
-     *  <p>Description of the Product displayed in search results.</p>
+     *  <p>Description of the Product as displayed in search engine results.</p>
      * @param metaDescription value to be set
      * @return Builder
      */
@@ -756,7 +757,8 @@ public class ProductDraftBuilder implements Builder<ProductDraft> {
 
     /**
      *  <p>User-defined unique identifier for the Product.</p>
-     *  <p>To update a Product using the Import API, the Product <code>key</code> must match the pattern <code>^[A-Za-z0-9_-]{2,256}$</code>.</p>
+     *  <p>This field is optional for backwards compatibility reasons, but we strongly recommend setting it. Keys are mandatory for importing Products with the Import API and the Merchant Center.</p>
+     *  <p>To update a Product using the Import API or Merchant Center, the Product <code>key</code> must match the pattern <code>^[A-Za-z0-9_-]{2,256}$</code>.</p>
      * @return key
      */
 
@@ -796,7 +798,7 @@ public class ProductDraftBuilder implements Builder<ProductDraft> {
     }
 
     /**
-     *  <p>Title of the Product displayed in search results.</p>
+     *  <p>Title of the Product as displayed in search engine results.</p>
      * @return metaTitle
      */
 
@@ -806,7 +808,7 @@ public class ProductDraftBuilder implements Builder<ProductDraft> {
     }
 
     /**
-     *  <p>Description of the Product displayed in search results.</p>
+     *  <p>Description of the Product as displayed in search engine results.</p>
      * @return metaDescription
      */