+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/product_tailoring/ProductTailoringDataImpl.java
@@ -79,7 +79,7 @@ public class ProductTailoringDataImpl implements ProductTailoringData, ModelBase
     }
 
     /**
-     *  <p>Tailored title of the Product used by external search engines for improved search engine performance.</p>
+     *  <p>Tailored title of the Product as displayed in search engine results.</p>
      */
 
     public com.commercetools.api.models.common.LocalizedString getMetaTitle() {
@@ -87,7 +87,7 @@ public class ProductTailoringDataImpl implements ProductTailoringData, ModelBase
     }
 
     /**
-     *  <p>Tailored description of the Product used by external search engines for improved search engine performance.</p>
+     *  <p>Tailored description of the Product as displayed in search engine results.</p>
      */
 
     public com.commercetools.api.models.common.LocalizedString getMetaDescription() {
@@ -95,7 +95,7 @@ public class ProductTailoringDataImpl implements ProductTailoringData, ModelBase
     }
 
     /**
-     *  <p>Tailored keywords related to the Product used by external search engines for improved search engine performance.</p>
+     *  <p>Tailored keywords related to the Product that are used by search engines.</p>
      */
 
     public com.commercetools.api.models.common.LocalizedString getMetaKeywords() {