+++ commercetools/commercetools-sdk-java-importapi/src/main/java-generated/com/commercetools/importapi/models/productvariants/ProductVariantImport.java
@@ -56,7 +56,10 @@ public interface ProductVariantImport extends ImportResource {
     public String getSku();
 
     /**
-     *  <p>Maps to <code>ProductVariant.isMasterVariant</code>.</p>
+     *  <ul>
+     *   <li>When creating a new ProductVariant, set to <code>false</code>; otherwise, the import operation will fail with a NewMasterVariantAdditionNotAllowed error.</li>
+     *   <li>Set to <code>true</code> if the ProductVariant exists and you want to set this ProductVariant as the Master Variant.</li>
+     *  </ul>
      * @return isMasterVariant
      */
     @NotNull
@@ -122,7 +125,10 @@ public interface ProductVariantImport extends ImportResource {
     public void setSku(final String sku);
 
     /**
-     *  <p>Maps to <code>ProductVariant.isMasterVariant</code>.</p>
+     *  <ul>
+     *   <li>When creating a new ProductVariant, set to <code>false</code>; otherwise, the import operation will fail with a NewMasterVariantAdditionNotAllowed error.</li>
+     *   <li>Set to <code>true</code> if the ProductVariant exists and you want to set this ProductVariant as the Master Variant.</li>
+     *  </ul>
      * @param isMasterVariant value to be set
      */