+++ changes.md
@@ -1,8 +1,24 @@
+**Api changes**
+
+<details>
+<summary>Added Property(s)</summary>
+
+- added property `source` to type `EventBridgeDestination`
+</details>
+
+
+<details>
+<summary>Added Method(s)</summary>
+
+- added method `apiRoot.withProjectKey().productTailoring().head()`
+</details>
+
 **Import changes**
 
 <details>
 <summary>Added Type(s)</summary>
 
 - added type `InvalidFieldsUpdateError`
+- added type `NewMasterVariantAdditionNotAllowedError`
 </details>